Increase clickable area for side nav links
Previously, it was a small 'a' in a big 'ul'. Now the 'a' takes up the full size.

ALSO, refactored nav link data in general to live in a 'nav_links_by_group'
helper, which is the master list of nav links.

This helps move some logic away from the templates so we can be a little more
precise about where and how nav links are rendered.